feat: add GroupEntity for dynamic group representation with proper article handling
